requirements.txt
cryptography>=3.4.0
web3>=5.23.0
py-solc-x>=1.1.1